requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
appium-python-client==2.2.0
    # via -r requirements.in
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.4.0
    # via
    #   outcome
    #   pytest
    #   trio
certifi==2021.10.8
    # via urllib3
cffi==1.15.0
    # via cryptography
cryptography==36.0.2
    # via
    #   pyopenssl
    #   urllib3
h11==0.13.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
outcome==1.1.0
    # via trio
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pyopenssl==22.0.0
    # via urllib3
pyparsing==3.0.7
    # via packaging
pysocks==1.7.1
    # via urllib3
pytest==7.1.1
    # via -r requirements.in
selenium==4.1.3
    # via appium-python-client
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
tomli==2.0.1
    # via pytest
trio==0.20.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure,socks]==1.26.9
    # via selenium
wsproto==1.1.0
    # via trio-websocket
